+++ private/dbutil/sqliteutil/query_test.go
@@ -6,7 +6,6 @@ package sqliteutil_test
 import (
 	"testing"
 
-	_ "github.com/jackc/pgx/v4/stdlib"
 	_ "github.com/mattn/go-sqlite3"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
