+++ private/dbutil/txutil/transactions.go
@@ -13,7 +13,7 @@ import (
 	"github.com/spacemonkeygo/monkit/v3"
 	"github.com/zeebo/errs"
 
-	"storj.io/storj/private/dbutil/pgutil"
+	"storj.io/storj/private/dbutil/pgutil/pgerrcode"
 	"storj.io/storj/private/tagsql"
 )
 
@@ -40,7 +40,7 @@ func WithTx(ctx context.Context, db tagsql.DB, txOpts *sql.TxOptions, fn func(co
 			if dur := time.Since(start); dur < 5*time.Minute && i < 10 {
 				// even though the resources (duration and count) allow us to issue a retry,
 				// we only should if the error claims we should.
-				if code := pgutil.ErrorCode(err); code == "CR000" || code == "40001" {
+				if code := pgerrcode.FromError(err); code == "CR000" || code == "40001" {
 					continue
 				}
 			} else {
